quality time refactor!
