+++ api/src/Service/QualityTime/PracticeTricksEvent.php
@@ -1,0 +1,56 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This file is part of the Poppy Seed Pets API.
+ *
+ * The Poppy Seed Pets API is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * The Poppy Seed Pets API is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with The Poppy Seed Pets API. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace App\Service\QualityTime;
+
+use App\Entity\Pet;
+use App\Entity\User;
+use App\Functions\ActivityHelpers;
+use App\Functions\ArrayFunctions;
+use App\Service\IRandom;
+
+class PracticeTricksEvent implements QualityTimeEvent
+{
+    public function __construct(
+        private readonly IRandom $rng,
+    )
+    {
+    }
+
+    public function isAvailable(User $user): bool
+    {
+        return true;
+    }
+
+    public function generate(User $user, array $pets): QualityTimeResult
+    {
+        $petNamesList = array_map(fn(Pet $pet) => ActivityHelpers::PetName($pet), $pets);
+        $everyonesNames = ArrayFunctions::list_nice([
+            ActivityHelpers::UserName($user, true),
+            ...$petNamesList
+        ]);
+
+        $randomPet = $this->rng->rngNextFromArray($petNamesList);
+
+        $tricks = [
+            "$randomPet finally mastered a perfect twirl!",
+            "$randomPet almost did a backflip, then burst into proud wiggles instead.",
+            "$randomPet learned to boop on command.",
+            "$randomPet showed everyone a new dance step.",
+        ];
+
+        $message = "$everyonesNames practiced tricks in the living room. " . $this->rng->rngNextFromArray($tricks);
+
+        return new QualityTimeResult($message, foodBased: false);
+    }
+}
